requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
Werkzeug==3.0.2
